+++ viteclient/src/pages/VillaAnna/VillaAnnaCountdownPage.jsx
@@ -2,7 +2,17 @@
 import { useState } from 'react';
 import axios from 'axios';
 import Layout from '../../Layout';
-import { FormControl, InputLabel, Select, MenuItem, Grid, Card, CardContent, CardHeader, Button } from '@mui/material';
+import {
+    FormControl,
+    InputLabel,
+    Select,
+    MenuItem,
+    Grid,
+    Card,
+    CardContent,
+    CardHeader,
+    Button
+} from '@mui/material';
 import { zoneOrder, bewaesserungsTopics } from '../../components/constants';
 import { HourField, MinuteField } from '../../components/index';
 
@@ -16,9 +26,14 @@ const VillaAnnacountdownPage = () => {
         setSelectedZone(event.target.value);
     }
 
-    const handleSendTopic = () => {
+    const handleSendTopic = (action) => {
         const selectedTopic = bewaesserungsTopics[zoneOrder.indexOf(selectedZone)];
-        axios.post(`${apiUrl}/countdown/setCountdown`, { topic: selectedTopic })
+        axios.post(`${apiUrl}/countdown/setCountdown`, {
+            topic: selectedTopic,
+            hours: selectedHour,
+            minutes: selectedMinute,
+            action: action
+        })
             .then(response => {
                 console.log('Response:', response.data);
             })
@@ -33,21 +48,23 @@ const VillaAnnacountdownPage = () => {
                 <Card>
                     <CardHeader title="Countdown einstellen" />
                     <CardContent>
-                        <FormControl fullWidth>
-                            <InputLabel id="zone-select-label">Zone</InputLabel>
-                            <Select
-                                labelId="zone-select-label"
-                                value={selectedZone}
-                                onChange={handleZoneChange}
-                            >
-                                {zoneOrder.map((zone, i) => (
-                                    <MenuItem value={zone} key={bewaesserungsTopics[i]}>
-                                        {zone}
-                                    </MenuItem>
-                                ))}
-                            </Select>
-                        </FormControl>
                         <Grid container spacing={2}>
+                            <Grid item xs={12}>
+                                <FormControl fullWidth>
+                                    <InputLabel id="zone-select-label">Zone</InputLabel>
+                                    <Select
+                                        labelId="zone-select-label"
+                                        value={selectedZone}
+                                        onChange={handleZoneChange}
+                                    >
+                                        {zoneOrder.map((zone, i) => (
+                                            <MenuItem value={zone} key={bewaesserungsTopics[i]}>
+                                                {zone}
+                                            </MenuItem>
+                                        ))}
+                                    </Select>
+                                </FormControl>
+                            </Grid>
                             <Grid item xs={6}>
                                 <HourField
                                     selectedHour={selectedHour}
@@ -60,11 +77,21 @@ const VillaAnnacountdownPage = () => {
                                     setSelectedMinute={setSelectedMinute}
                                 />
                             </Grid>
+                            <Grid item xs={4}>
+                                <Button variant='contained' color='primary' fullWidth onClick={() => handleSendTopic('start')}>
+                                    Start
+                                </Button>
                             </Grid>
-                        <Grid item xs={12}>
-                            <Button variant='contained' color='primary' fullWidth onClick={handleSendTopic}>
-                                Start
-                            </Button>
+                            <Grid item xs={4}>
+                                <Button variant='contained' color='info' fullWidth onClick={() => handleSendTopic('stop')}>
+                                    Stop
+                                </Button>
+                            </Grid>
+                            <Grid item xs={4}>
+                                <Button variant='contained' color='warning' fullWidth onClick={() => handleSendTopic('reset')}>
+                                    Reset
+                                </Button>
+                            </Grid>
                         </Grid>
                     </CardContent>
                 </Card>